+++ src/Form/LegacyConsumer/Commands/HookCommandInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Give\Form\LegacyConsumer\Commands;
+
+interface HookCommandInterface {
+	public function __invoke( $hook );
+}