+++ src/Form/LegacyConsumer/templates/label.html.php
@@ -1,0 +1,6 @@
+<label class="give-label" for="give-<?php echo $field->getName(); ?>">
+	<?php echo $field->getLabel(); ?>
+	<?php if ( $field->isRequired() ) : ?>
+	&nbsp;<span class="give-required-indicator">*</span>
+	<?php endif; ?>
+</label>