+++ src/Framework/FieldsAPI/FormField/FieldHelpText.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Give\Framework\FieldsAPI\FormField;
+
+trait FieldHelpText {
+
+	/** @var string */
+	protected $helpText;
+
+	/**
+	 * @param string $helpText
+	 * @return $this
+	 */
+	public function helpText( $helpText ) {
+		$this->helpText = $helpText;
+		return $this;
+	}
+
+	public function getHelpText() {
+		return $this->helpText;
+	}
+}