+++ src/Framework/FieldsAPI/FormField/FieldStoreAsMeta.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace Give\Framework\FieldsAPI\FormField;
+
+/**
+ * @unreleased
+ */
+trait FieldStoreAsMeta {
+
+	/**
+	 * @unreleased
+	 * @var bool
+	 */
+	protected $storeAsDonorMeta = false;
+
+	/**
+	 * @unreleased
+	 * @return $this
+	 */
+	public function storeAsDonorMeta() {
+		$this->storeAsDonorMeta = true;
+		return $this;
+	}
+
+	/**
+	 * @unreleased
+	 * @return bool
+	 */
+	public function shouldStoreAsDonorMeta() {
+		return $this->storeAsDonorMeta;
+	}
+}