+++ src/Framework/FieldsAPI/FormFieldMediator/Contract.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Give\Framework\FieldsAPI\FieldMediator;
+
+use Give\Framework\FieldsAPI\Field;
+
+interface Contract {
+	public function __invoke( Field $field );
+}